Исправлено копирование информации при удалении пользователя

+++ scripts/deleteAccount.php
@@ -14,24 +14,24 @@
 			header("Location: ../index.php");
 		}
 	}
-	else
-	{
-		$userResult = $mysqli->query("SELECT * FROM users WHERE id = '".$_SESSION['userID']."'");
+	else {
+		$userResult = $mysqli->query("SELECT * FROM users WHERE id = '" . $_SESSION['userID'] . "'");
 		$user = $userResult->fetch_assoc();
 
-		$mysqli->query("INSERT INTO users_deleted (id, email, organisation, person, phone) VALUES ('".$user['id']."', '".$user['email']."', '".$user['organisation']."', '".$user['phone']."')");
-
-		if($mysqli->query("DELETE FROM users WHERE id = '".$_SESSION['userID']."'"))
-		{
-			unset($_SESSION['userID']);
-			$_SESSION['delete'] = 'ok';
-			header("Location: ../index.php");
-		}
-		else
-		{
+		if ($mysqli->query("INSERT INTO users_deleted (id, email, organisation, person, phone) VALUES ('" . $user['id'] . "', '" . $user['email'] . "', '" . $user['organisation'] . "', '" . $user['person'] . "' ,'" . $user['phone'] . "')")) {
+			if($mysqli->query("DELETE FROM users WHERE id = '".$_SESSION['userID']."'"))
+			{
+				unset($_SESSION['userID']);
+				$_SESSION['delete'] = 'ok';
+				header("Location: ../index.php");
+			}
+			else
+			{
+				$_SESSION['delete'] = 'failed';
+				header("Location: ../settings.php?s=3");
+			}
+		} else {
 			$_SESSION['delete'] = 'failed';
-			header("Location: ../settings.php?s=3");	
+			header("Location: ../settings.php?s=3");
 		}
 	}
-
-?>
